Move execute_sql utility function into db module
