+++ back/test/model_tests/test_utils.py
@@ -7,7 +7,6 @@ from boxtribute_server.db import DatabaseManager
 from boxtribute_server.models.definitions.base import Base
 from boxtribute_server.models.definitions.history import DbChangeHistory
 from boxtribute_server.models.utils import (
-    execute_sql,
     format_sql,
     handle_non_existing_resource,
 )
@@ -48,5 +47,5 @@ def test_format_sql():
         == 'SELECT IF((`t1`.`changedate` > 0), "new", NULL) AS `x` FROM `history` AS `t1` WHERE (((`t1`.`tablename` = "stock") AND (`t1`.`changes` = "product_id")) AND (`t1`.`from_int` = 1))'  # noqa
     )
     # Validate the formatted SQL
-    result = execute_sql(query=formatted_query, database=Base._meta.database)
-    assert result == []
+    result = Base._meta.database.execute_sql(formatted_query).fetchall()
+    assert result == ()
